204. Count Primes
