+++ src/test/kotlin/me/bossm0n5t3r/leetcode/countPrimes/CountPrimesTest.kt
@@ -1,10 +1,10 @@
-package countPrimes
+package me.bossm0n5t3r.leetcode.countPrimes
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class CountPrimesTest {
-    private val countPrimes = CountPrimes()
+    private val countPrimes = CountPrimes.Solution()
 
     data class CountPrimesTestData(
         val n: Int,
